code refactor according to checkstyle rules
